requirements.txt
appnope==0.1.0
bleach==2.1.1
certifi==2017.11.5
chardet==3.0.4
decorator==4.1.2
entrypoints==0.2.3
future==0.16.0
h5py==2.7.1
html5lib==1.0b10
idna==2.6
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.5
jedi==0.11.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.1
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.2.5
notebook==5.2.1
numpy==1.13.3
oauthlib==2.0.6
pandas==0.21.0
pandocfilters==1.4.2
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyaes==1.6.1
pyasn1==0.4.2
Pygments==2.2.0
python-dateutil==2.6.1
python-telegram-bot==8.1.1
pytz==2017.3
pyzmq==16.0.3
qtconsole==4.3.1
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
simplegeneric==0.8.1
six==1.11.0
Telethon==0.15.5.3
terminado==0.8
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
tweepy==3.5.0
tweetfeels==0.4.0
twython==3.6.0
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.8
